+++ src/main/java/utils/UserUtils.java
@@ -1,0 +1,8 @@
+package utils;
+
+import lombok.AccessLevel;
+import lombok.NoArgsConstructor;
+
+@NoArgsConstructor(access = AccessLevel.PRIVATE)
+public class UserUtils {
+}
